feat: Add PermissionGuard component and integrate into other components
This commit adds the `PermissionGuard` component, which allows for conditional rendering of child components based on user permissions. It uses the `useSelector` hook from "react-redux" to access user information and check for the required permission.

Additionally, this component is integrated into other components to provide permission-based rendering. The `PermissionGuard` component is used to control the visibility of specific parts of the application based on the user's permissions.